+++ src/basic/components/index.tsx
@@ -1,8 +1,8 @@
-export * from './Header';
-export * from './Notification';
 export * from './main/Coupon';
 export * from './main/Product';
 export * from './main/Cart';
 export * from './main/Payment';
 export * from './admin/AdminPage';
 export * from './main/MainPage';
+export * from './Header';
+export * from './Notification';
