+++ src/basic/components/main/index.tsx
@@ -1,0 +1,4 @@
+export * from './Cart';
+export * from './Coupon';
+export * from './Payment';
+export * from './Product';
