+++ src/basic/hooks/index.ts
@@ -1,5 +1,6 @@
 export * from './useDebounce';
 export * from './useNotification';
+export * from './useLocalStorage';
 export * from './useProduct';
 export * from './useCart';
 export * from './useCoupon';
